Process notifications in the sites task
Notifications are needed not only for dirty log, but also for sites
online status tracking.
